maxMinimizeSteps: make it 64 bits
As we might hit 4G conflicts during a maxsat run, make sure we limit
using 64 bits. As we need a way to tell it's disabled, actually use
only 63 bits, and use the sign to disable. By default, the limit is
-1, hence turned off.

+++ include/MaxSATSolver.h
@@ -238,7 +238,7 @@ public:
                               uint64_t &cost,
                               uint64_t maxCost = UINT64_MAX,
                               const std::vector<int> *startAssignment = 0,
-                              int maxMinimizeSteps = -1);
+                              int64_t maxMinimizeSteps = -1);
 };
 
 #endif
