+++ smax-src/SMaxSolver.h
@@ -143,7 +143,7 @@ protected:
                               uint64_t &cost,
                               uint64_t maxCost = INT64_MAX,
                               const std::vector<int> *startAssignment = 0,
-                              int maxMinimizeSteps = -1);
+                              int64_t maxMinimizeSteps = -1);
 
     int getValue(unsigned variable)
     {
